Commit before pull
